Added logs for debugging and: simplified files to make it debugging easier
