+++ app/notification/application/map_contents.py
@@ -1,8 +1,6 @@
 from __future__ import annotations
 
 import collections
-import json
-import uuid
 from dataclasses import dataclass
 from datetime import datetime
 from io import BytesIO
@@ -10,13 +8,14 @@ from io import StringIO
 from typing import TYPE_CHECKING
 
 import pytz
-from app.notification.model.multi_input_data import MultiInput
+from app.notification.application.application_utils import (
+    sort_questions_and_answers,
+)
 from app.notification.model.notification_utils import format_answer
 from app.notification.model.notification_utils import simplify_title
 from app.notification.notification_contents_base_class import (
     _NotificationContents,
 )
-from bs4 import BeautifulSoup
 from config import Config
 from flask import current_app
 from fsd_utils.config.notify_constants import NotifyConstants
@@ -110,49 +109,9 @@ class Application(_NotificationContents):
         questions_answers = collections.defaultdict(dict)
         forms = cls.get_forms(notification)
         form_names = cls.get_form_names(notification)
-
-        for form_name in form_names:
-            for form in forms:
-                if form_name in form[NotifyConstants.APPLICATION_NAME_FIELD]:
-                    for question in form[
-                        NotifyConstants.APPLICATION_QUESTIONS_FIELD
-                    ]:
-                        for field in question["fields"]:
-                            answer = field.get("answer")
-                            clean_html_answer = cls.remove_html_tags(answer)
-
-                            if field["type"] == "file":
-                                # we check if the question type is "file"
-                                # then we remove the aws
-                                # key attached to the answer
-
-                                if isinstance(clean_html_answer, str):
-                                    questions_answers[form_name][
-                                        field["title"]
-                                    ] = clean_html_answer.split("/")[-1]
-                                else:
-                                    questions_answers[form_name][
-                                        field["title"]
-                                    ] = clean_html_answer
-                            elif (
-                                isinstance(clean_html_answer, bool)
-                                and field["type"] == "list"
-                            ):
-                                questions_answers[form_name][
-                                    field["title"]
-                                ] = ("Yes" if clean_html_answer else "No")
-                            elif (
-                                isinstance(clean_html_answer, list)
-                                and field["type"] == "multiInput"
-                            ):
-
-                                questions_answers[form_name][
-                                    field["title"]
-                                ] = cls.map_multi_input_data(clean_html_answer)
-                            else:
-                                questions_answers[form_name][
-                                    field["title"]
-                                ] = clean_html_answer
+        questions_answers = sort_questions_and_answers(
+            forms, form_names, questions_answers
+        )
         return questions_answers
 
     @classmethod
@@ -199,95 +158,4 @@ class Application(_NotificationContents):
         bytes_object = BytesIO(convert_to_bytes)
         return bytes_object
 
-    @classmethod
-    def remove_html_tags(cls, answer):
-        """
-        Removes HTML tags from the provided answer and returns the cleaned text.
-
-        Args:
-            answer (str): The answer containing HTML tags.
-
-        Returns:
-            str: The cleaned text with HTML tags removed.
-        """
-
-        try:
-
-            if answer is None or isinstance(answer, (bool, list)):
-                return answer
-
-            # Check if answer looks like a URL
-            if answer.startswith("http://") or answer.startswith("https://"):
-                return answer
-
-            soup = BeautifulSoup(answer, "html.parser")
-            indent = " " * 5
-
-            if not soup.find():
-                return answer.strip()
-
-            if not (soup.ul or soup.ol):
-                # Handle other HTML tags
-                cleaned_text = soup.get_text()
-                cleaned_text = cleaned_text.replace("\xa0", "")
-                return cleaned_text.strip()
-
-            soup_list = soup.ul or soup.ol
-            list_items = []
-            for index, li in enumerate(soup_list.find_all("li"), start=1):
-                separator = "." if soup.ul else f"{index}."
-                if li.get_text() == "\xa0":
-                    continue
-
-                if index > 1:
-                    list_items.append(f"{indent}{separator} {li.get_text()}")
-                else:
-                    list_items.append(f"{separator} {li.get_text()}")
-            return "\n".join(list_items)
-
-        except Exception as e:
-            current_app.logger.error(
-                f"Error occurred while processing HTML tag: {answer}", e
-            )
-            return answer
-
-    @classmethod
-    def map_multi_input_data(cls, multi_input_data):
-        """
-        Map the multi-input data to a sorted dictionary and process it.
-
-        Args:
-            multi_input_data (list): The list of dictionaries representing the multi-input data.
-
-        Returns:
-            str: The processed output as a formatted string.
-        """
-
-        key = None
-        value = None
-        sorted_data = {}
-        for item in multi_input_data:
-            if len(item) < 2:
-                for value in item.values():
-                    key = str(uuid.uuid4())
-                    sorted_data[key] = value
-            else:
-                try:
-                    key, *values = item.values()
-                    sorted_data[key] = values
-                except TypeError:
-                    value = tuple(item.values())
-                    if isinstance(value[0], dict) and isinstance(
-                        value[1], str
-                    ):
-                        *values, key = item.values()
-                        sorted_data[key] = values
-                except:  # noqa
-                    current_app.logger.error(
-                        "Could not format the data correctly for"
-                        f" {multi_input_data}"
-                    )
-                    key, *values = item.values()
-                    sorted_data[json.dumps(key)] = values
-        output = MultiInput.process_data(sorted_data)
-        return "\n".join(output)
+
